Add: Multiple saf inputs
The script now accepts multiple input saf files and can process
them all together, similar to how an individual saf was processed
before. Added in an additional switch to average over the number of
saf inputs. So the script can produce either the total number of
events or a scales cross section.