Add: event weight rescaling
Modify histogram to take event weights as an argument, one is then
able to pass event weights rather than xsec and nevents which is
necessary for NLO events with no weight information, as weights are
not simple xsec/nevents.

To this end change the main method of the script so that it chooses
between the extended and the usual methods based on user choices.